Move WorkspaceMigration code to src/workspaces/migration.
