+++ src/pages/workspaces/WorkspaceMigration.ts
@@ -1,0 +1,10 @@
+import { WorkspaceMigrationPage } from 'src/workspaces/migration/WorkspaceMigrationPage';
+
+export const navPaths = [
+  {
+    name: 'workspace-migration',
+    path: '/workspace-migration',
+    component: WorkspaceMigrationPage,
+    title: 'Workspace Migration',
+  },
+];
