+++ src/workspaces/migration/BillingProjectList.test.ts
@@ -2,9 +2,9 @@ import { abandonedPromise } from '@terra-ui-packages/core-utils';
 import { act, screen, within } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
 import { Ajax } from 'src/libs/ajax';
-import { BillingProjectList, inProgressRefreshRate } from 'src/pages/workspaces/migration/BillingProjectList';
-import { mockServerData } from 'src/pages/workspaces/migration/migration-utils.test';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
+import { BillingProjectList, inProgressRefreshRate } from 'src/workspaces/migration/BillingProjectList';
+import { mockServerData } from 'src/workspaces/migration/migration-utils.test';
 
 type AjaxContract = ReturnType<typeof Ajax>;
 type AjaxWorkspacesContract = AjaxContract['Workspaces'];
